♻️ refactor: refactor the code
