+++ src/utils/fetch/__tests__/index.test.ts
@@ -1,46 +1,13 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { ZodError } from 'zod';
 
-import { ErrorResponse } from '@/types/fetch';
-
-import { FetchEventSourceInit } from './fetchEventSource';
-import { fetchEventSource } from './fetchEventSource';
-import { fetchSSE, getMessageError, parseToolCalls } from './index';
+import { FetchEventSourceInit } from '../fetchEventSource';
+import { fetchEventSource } from '../fetchEventSource';
+import { fetchSSE } from '../fetchSSE';
 
 // 模拟 i18next
 vi.mock('i18next', () => ({
   t: vi.fn((key) => `translated_${key}`),
 }));
-
-// 模拟 Response
-const createMockResponse = (body: any, ok: boolean, status: number = 200) => ({
-  ok,
-  status,
-  json: vi.fn(async () => body),
-  clone: vi.fn(function () {
-    // @ts-ignore
-    return this;
-  }),
-  text: vi.fn(async () => JSON.stringify(body)),
-  body: {
-    getReader: () => {
-      let done = false;
-      return {
-        read: () => {
-          if (!done) {
-            done = true;
-            return Promise.resolve({
-              value: new TextEncoder().encode(JSON.stringify(body)),
-              done: false,
-            });
-          } else {
-            return Promise.resolve({ done: true });
-          }
-        },
-      };
-    },
-  },
-});
 
 vi.mock('./fetchEventSource', () => ({
   fetchEventSource: vi.fn(),
@@ -49,169 +16,6 @@ vi.mock('./fetchEventSource', () => ({
 // 在每次测试后清理所有模拟
 afterEach(() => {
   vi.restoreAllMocks();
-});
-
-describe('getMessageError', () => {
-  it('should handle business error correctly', async () => {
-    const mockErrorResponse: ErrorResponse = {
-      body: 'Error occurred',
-      errorType: 'InvalidAccessCode',
-    };
-    const mockResponse = createMockResponse(mockErrorResponse, false, 400);
-
-    const error = await getMessageError(mockResponse as any);
-
-    expect(error).toEqual({
-      body: mockErrorResponse.body,
-      message: 'translated_response.InvalidAccessCode',
-      type: mockErrorResponse.errorType,
-    });
-    expect(mockResponse.json).toHaveBeenCalled();
-  });
-
-  it('should handle regular error correctly', async () => {
-    const mockResponse = createMockResponse({}, false, 500);
-    mockResponse.json.mockImplementationOnce(() => {
-      throw new Error('Failed to parse');
-    });
-
-    const error = await getMessageError(mockResponse as any);
-
-    expect(error).toEqual({
-      message: 'translated_response.500',
-      type: 500,
-    });
-    expect(mockResponse.json).toHaveBeenCalled();
-  });
-
-  it('should handle timeout error correctly', async () => {
-    const mockResponse = createMockResponse(undefined, false, 504);
-    const error = await getMessageError(mockResponse as any);
-
-    expect(error).toEqual({
-      message: 'translated_response.504',
-      type: 504,
-    });
-  });
-});
-
-describe('parseToolCalls', () => {
-  it('should create add new item', () => {
-    const chunk = [
-      { index: 0, id: '1', type: 'function', function: { name: 'func', arguments: '' } },
-    ];
-
-    const result = parseToolCalls([], chunk);
-    expect(result).toEqual([
-      { id: '1', type: 'function', function: { name: 'func', arguments: '' } },
-    ]);
-  });
-
-  it('should update arguments if there is a toolCall', () => {
-    const origin = [{ id: '1', type: 'function', function: { name: 'func', arguments: '' } }];
-
-    const chunk1 = [{ index: 0, function: { arguments: '{"lo' } }];
-
-    const result1 = parseToolCalls(origin, chunk1);
-    expect(result1).toEqual([
-      { id: '1', type: 'function', function: { name: 'func', arguments: '{"lo' } },
-    ]);
-
-    const chunk2 = [{ index: 0, function: { arguments: 'cation\\": \\"Hangzhou\\"}' } }];
-    const result2 = parseToolCalls(result1, chunk2);
-
-    expect(result2).toEqual([
-      {
-        id: '1',
-        type: 'function',
-        function: { name: 'func', arguments: '{"location\\": \\"Hangzhou\\"}' },
-      },
-    ]);
-  });
-
-  it('should add a new tool call if the index is different', () => {
-    const origin = [
-      {
-        id: '1',
-        type: 'function',
-        function: { name: 'func', arguments: '{"location\\": \\"Hangzhou\\"}' },
-      },
-    ];
-
-    const chunk = [
-      {
-        index: 1,
-        id: '2',
-        type: 'function',
-        function: { name: 'func', arguments: '' },
-      },
-    ];
-
-    const result1 = parseToolCalls(origin, chunk);
-    expect(result1).toEqual([
-      {
-        id: '1',
-        type: 'function',
-        function: { name: 'func', arguments: '{"location\\": \\"Hangzhou\\"}' },
-      },
-      { id: '2', type: 'function', function: { name: 'func', arguments: '' } },
-    ]);
-  });
-
-  it('should update correct arguments if there are multi tool calls', () => {
-    const origin = [
-      {
-        id: '1',
-        type: 'function',
-        function: { name: 'func', arguments: '{"location\\": \\"Hangzhou\\"}' },
-      },
-      { id: '2', type: 'function', function: { name: 'func', arguments: '' } },
-    ];
-
-    const chunk = [{ index: 1, function: { arguments: '{"location\\": \\"Beijing\\"}' } }];
-
-    const result1 = parseToolCalls(origin, chunk);
-    expect(result1).toEqual([
-      {
-        id: '1',
-        type: 'function',
-        function: { name: 'func', arguments: '{"location\\": \\"Hangzhou\\"}' },
-      },
-      {
-        id: '2',
-        type: 'function',
-        function: { name: 'func', arguments: '{"location\\": \\"Beijing\\"}' },
-      },
-    ]);
-  });
-
-  it('should throw error if incomplete tool calls data', () => {
-    const origin = [
-      {
-        id: '1',
-        type: 'function',
-        function: { name: 'func', arguments: '{"location\\": \\"Hangzhou\\"}' },
-      },
-    ];
-
-    const chunk = [{ index: 1, id: '2', type: 'function' }];
-
-    try {
-      parseToolCalls(origin, chunk as any);
-    } catch (e) {
-      expect(e).toEqual(
-        new ZodError([
-          {
-            code: 'invalid_type',
-            expected: 'object',
-            received: 'undefined',
-            path: ['function'],
-            message: 'Required',
-          },
-        ]),
-      );
-    }
-  });
 });
 
 describe('fetchSSE', () => {
