+++ src/utils/fetch/parseError.ts
@@ -1,0 +1,26 @@
+import { t } from 'i18next';
+
+import { ErrorResponse, ErrorType } from '@/types/fetch';
+import { ChatMessageError } from '@/types/message';
+
+export const getMessageError = async (response: Response) => {
+  let chatMessageError: ChatMessageError;
+
+  // 尝试取一波业务错误语义
+  try {
+    const data = (await response.json()) as ErrorResponse;
+    chatMessageError = {
+      body: data.body,
+      message: t(`response.${data.errorType}` as any, { ns: 'error' }),
+      type: data.errorType,
+    };
+  } catch {
+    // 如果无法正常返回，说明是常规报错
+    chatMessageError = {
+      message: t(`response.${response.status}` as any, { ns: 'error' }),
+      type: response.status as ErrorType,
+    };
+  }
+
+  return chatMessageError;
+};
